\name{CCTable11.1a}
\alias{CCTable11.1a}
\title{Cochran and Cox design}
\usage{
	data(CCTable11.1a)
}

\description{
	A blocked experimental design from Cochran and Cox (1957).

	A Partially balanced incomplete block in 3 reps, and 27 blocks.
}

\source{
	Cochran, W.G. and Cox, G.M. (1957). \emph{Experimental designs}. Wiley, N.Y.
}

\keyword{datasets}

