\name{GVTable1}
\alias{GVTable1}
\title{Goos Vandebroek Table 1}
\usage{
	data(GVTable1)
}

\description{
	A blocked experimental design from Goos and Vandebroek (2003).

	A matrix giving one whole plot and four within plot variables, 
		in 21 blocks of 2.
	The same as TGTable5, but produced by Goos Vandoebroek algorithm.
}

\source{
	Goos, P. and Vandebroek, M. (2003). D-optimal split-plot designs with given numbers and 
		sizes of whole plots. \emph{Technometrics.} 45-3. 235-245.
}

\keyword{datasets}

