\name{GVTable3}
\alias{GVTable3}
\title{Goos Vandebroek Table 3}
\usage{
	data(GVTable3)
}

\description{
	A blocked experimental design from Goos and Vandebroek (2003).

	A matrix giving one whole and two within plot variables, 9 blocks of 21.

}

\source{
	Goos, P. and Vandebroek, M. (2003). D-optimal split-plot designs with given numbers and 
		sizes of whole plots. \emph{Technometrics.} 45-3. 235-245.
}

\keyword{datasets}


