\name{model.matrix.terms}
\alias{model.matrix.terms}
\title{Internal model.matrix.terms  function}
\description{
  defines a model.matrix function for the terms class
}
\usage{
\method{model.matrix}{terms}(...) 
}
\details{
  This is not to be called by the user.
  It fixes a bug whose solution was suggested  by Bill Dunlap, TIBCO Software Inc.
}
\keyword{design}
\keyword{internal}
