\name{TGTable3}
\alias{TGTable3}
\title{Trinca Gilmour Table 3}
\usage{
	data(TGTable3)
}

\description{
	A blocked experimental design from Trinca and Gilmour (2001).

	A data.frame giving four three level variable in five blocks. There
		are 45 trials in all.
}

\source{
	Trinca, L.A. and Gilmour, S.G. (2001). Multistratum response surface designs. 
		\emph{Technometrics}. 43-1. 25-33.
}

\keyword{datasets}

