\name{TGTable5}
\alias{TGTable5}
\title{Trinca Gilmour Table 5}
\usage{
	data(TGTable5)
}

\description{
	A blocked experimental design from Trinca and Gilmour (2001).

	A matrix giving one whole plot and four within plot variables, 
		in 21 blocks of 2.
}

\source{
	Trinca, L.A. and Gilmour, S.G. (2001). Multistratum response surface designs. 
		\emph{Technometrics}. 43-1. 25-33.
}

\keyword{datasets}

